+++ inlaprog/src/inla-likelihood.c
@@ -597,7 +597,7 @@ int inla_read_data_likelihood(inla_tp *mb, dictionary *UNUSED(ini), int UNUSED(s
 
 	case L_BINOMIALMIX:
 	{
-		assert(ncol_data_all == 16);
+		assert(ncol_data_all <= BINOMIALMIX_NBETA + 7);
 		idiv = ncol_data_all;
 		na = ncol_data_all - 2;
 		for (i = 0; i < na; i++) {
@@ -714,6 +714,7 @@ int inla_read_data_likelihood(inla_tp *mb, dictionary *UNUSED(ini), int UNUSED(s
 	if (ds->data_id == L_BINOMIALMIX) {
 		// re-arrange data so they become more accessible
 		ds->data_observations.binmix_dat = Calloc(mb->predictor_ndata, double *);
+		ds->data_observations.binmix_m = (int) (na - 5) / 2L;
 		for (i = 0; i < mb->predictor_ndata; i++) {
 			ds->data_observations.binmix_dat[i] = Calloc(na, double);
 			for (k = 0; k < na; k++) {
@@ -2981,6 +2982,26 @@ int loglikelihood_occupancy(int thread_id, double *__restrict logll, double *__r
 		return GMRFLib_SUCCESS;
 	}
 
+#if 0
+	// debug-code to check the likelihood
+	if (m != 99) {
+		double xx[99];
+		for(int i = 0; i < 99; i++) {
+			xx[i] = -4.0 + 8.0 * i/99.0;
+		}
+		double llxx[99];
+		loglikelihood_occupancy(thread_id, llxx, xx, 99, idx, NULL, NULL, arg, NULL);
+
+		char *fnm;
+		GMRFLib_sprintf(&fnm, "FILE%1d.txt", idx);
+		FILE *fp = fopen(fnm, "w");
+		for(int i = 0; i < 99; i++) {
+			fprintf(fp,  "%f %f\n", xx[i], llxx[i]);
+		}
+		fclose(fp);
+	}
+#endif
+	
 	Data_section_tp *ds = (Data_section_tp *) arg;
 	int ny = ds->data_observations.occ_ny[idx];
 
@@ -2990,9 +3011,6 @@ int loglikelihood_occupancy(int thread_id, double *__restrict logll, double *__r
 	}
 
 	LINK_INIT;
-
-	// code is hard-coded for this case... more work is needed to make it general
-	assert(PREDICTOR_LINK_EQ(link_logit));
 
 	int nb = ds->data_observations.occ_nbeta;
 	int yzero = ds->data_observations.occ_yzero[idx];
@@ -3009,7 +3027,7 @@ int loglikelihood_occupancy(int thread_id, double *__restrict logll, double *__r
 		const int mkl_lim = 4L;
 		double logll0 = 0.0;
 
-		if (ds->data_observations.link_simple_invlinkfunc == link_logit) {
+		if (PREDICTOR_SIMPLE_LINK_EQ(link_logit)) {
 			if (ny >= mkl_lim) {
 				size_t mm = GMRFLib_align_simple((size_t) ny, sizeof(double));
 				double w[mm], ww[mm];
@@ -3037,69 +3055,104 @@ int loglikelihood_occupancy(int thread_id, double *__restrict logll, double *__r
 			}
 		}
 
-		double x_critical = -0.5 * logll0;
+		double x_critical = 0.0;
+		if (PREDICTOR_LINK_EQ(link_logit)) {
+			x_critical = -0.5 * logll0;
+		} else if (PREDICTOR_LINK_EQ(link_cloglog)) {
+			double a = exp(logll0);
+			x_critical = log(my_lambert_W0((1.0 - a) / (a * exp(1.0))) + 1.0);
+		} else {
+			assert(0 == 1);
+		}
 		double x0 = 0.900 * x_critical;
 		double x1 = 0.999 * x_critical;
 		int tail = (GMRFLib_max_value(x, m, NULL) + off > x0);
 
 		if (!tail && PREDICTOR_SCALE == 1.0) {
 			if (yzero) {
-				double elogll0 = exp(logll0);
-
-				if (m >= mkl_lim) {
-					size_t mm = GMRFLib_align_simple((size_t) m, sizeof(double));
-					double xx[mm], exx[mm];
+				if (PREDICTOR_LINK_EQ(link_logit)) {
+					double elogll0 = exp(logll0);
+					if (m >= mkl_lim) {
+						size_t mm = GMRFLib_align_simple((size_t) m, sizeof(double));
+						double xx[mm], exx[mm];
 #pragma omp simd
-					for (int i = 0; i < m; i++) {
-						xx[i] = x[i] + off;
+						for (int i = 0; i < m; i++) {
+							xx[i] = x[i] + off;
+						}
+						GMRFLib_exp(m, xx, exx);
+#pragma omp simd
+						for (int i = 0; i < m; i++) {
+							double a = 1.0 + 1.0 / exx[i];
+							double b = (1.0 + exx[i]) * elogll0;
+							xx[i] = 1.0 / a + 1.0 / b;
+						}
+						GMRFLib_log(m, xx, exx);
+#pragma omp simd
+						for (int i = 0; i < m; i++) {
+							logll[i] = logll0 + exx[i];
+						}
+					} else {
+#pragma omp simd
+						for (int i = 0; i < m; i++) {
+							double ex = exp(x[i] + off);
+							double exd = 1.0 / ex;
+							logll[i] = logll0 + log(1.0 / (1.0 + exd) + 1.0 / ((1.0 + ex) * elogll0));
+						}
 					}
-					GMRFLib_exp(m, xx, exx);
+				} else if (PREDICTOR_LINK_EQ(link_cloglog)) {
 #pragma omp simd
 					for (int i = 0; i < m; i++) {
-						double a = 1.0 + 1.0 / exx[i];
-						double b = (1.0 + exx[i]) * elogll0;
-						xx[i] = 1.0 / a + 1.0 / b;
-					}
-					GMRFLib_log(m, xx, exx);
-#pragma omp simd
-					for (int i = 0; i < m; i++) {
-						logll[i] = logll0 + exx[i];
+						double log_phi = 0.0, log_1m_phi = 0.0;
+						link_log_invcloglog2(x[i] + off, &log_phi, &log_1m_phi);
+						logll[i] = GMRFLib_logsum(logll0 + log_phi, log_1m_phi);
 					}
 				} else {
 #pragma omp simd
 					for (int i = 0; i < m; i++) {
-						double ex = exp(x[i] + off);
-						double exd = 1.0 / ex;
-						logll[i] = logll0 + log(1.0 / (1.0 + exd) + 1.0 / ((1.0 + ex) * elogll0));
+						double phi = PREDICTOR_INVERSE_LINK(x[i] + off);
+						logll[i] = GMRFLib_logsum(logll0 + LOG_p(phi), LOG_1mp(phi));
 					}
 				}
 			} else {
-				if (m >= mkl_lim) {
-					size_t mm = GMRFLib_align_simple((size_t) m, sizeof(double));
-					double w[mm], ww[mm];
+				if (PREDICTOR_LINK_EQ(link_logit)) {
+					if (m >= mkl_lim) {
+						size_t mm = GMRFLib_align_simple((size_t) m, sizeof(double));
+						double w[mm], ww[mm];
 #pragma omp simd
-					for (int i = 0; i < m; i++) {
-						w[i] = -(x[i] + off);
+						for (int i = 0; i < m; i++) {
+							w[i] = -(x[i] + off);
+						}
+						GMRFLib_exp(m, w, ww);
+						GMRFLib_log1p(m, ww, w);
+#pragma omp simd
+						for (int i = 0; i < m; i++) {
+							logll[i] = logll0 - w[i];
+						}
+					} else {
+#pragma omp simd
+						for (int i = 0; i < m; i++) {
+							double exd = exp(-(x[i] + off));
+							logll[i] = logll0 - log1p(exd);
+						}
 					}
-					GMRFLib_exp(m, w, ww);
-					GMRFLib_log1p(m, ww, w);
+				} else if (PREDICTOR_LINK_EQ(link_cloglog)) {
 #pragma omp simd
 					for (int i = 0; i < m; i++) {
-						logll[i] = logll0 - w[i];
+						logll[i] = logll0 + link_log_invcloglog(x[i] + off);
 					}
 				} else {
 #pragma omp simd
 					for (int i = 0; i < m; i++) {
-						double exd = exp(-(x[i] + off));
-						logll[i] = logll0 - log1p(exd);
+						double phi = PREDICTOR_INVERSE_LINK_NO_SCALE(x[i] + off);
+						logll[i] = logll0 + log(phi);
 					}
 				}
 			}
 		} else {
 			if (yzero) {
-				if (tail) {
-					// we fix the tail in the low-likelihood area to avoid issues
-					// this is documented in 'internal-doc/occupancy/description.pdf'
+				// we fix the tail in the low-likelihood area to avoid issues
+				// this is documented in 'internal-doc/occupancy/description.pdf'
+				if (PREDICTOR_LINK_EQ(link_logit)) {
 					double a = exp(logll0);
 					for (int i = 0; i < m; i++) {
 						double xoff = x[i] + off;
@@ -3120,6 +3173,39 @@ int loglikelihood_occupancy(int thread_id, double *__restrict logll, double *__r
 							logll[i] = c0 + dx * (c1 + 0.5 * c2 * dx);
 						}
 					}
+				} else if (PREDICTOR_LINK_EQ(link_cloglog)) {
+					double a = exp(logll0);
+					for (int i = 0; i < m; i++) {
+						double xoff = x[i] + off;
+						double xx = PREDICTOR_INVERSE_IDENTITY_LINK(xoff);
+						if (xx <= x0) {
+							double log_phi = 0.0, log_1m_phi = 0.0;
+							link_log_invcloglog2(xx, &log_phi, &log_1m_phi);
+							logll[i] = GMRFLib_logsum(logll0 + log_phi, log_1m_phi);
+						} else {
+							double z = (xx - x0) / (x1 - x0);
+							double xx0 = x0 + (x1 - x0) * z / (1.0 + z);
+							double dx = xx - xx0;
+
+							double cg = a;
+							double t1 = exp(xx0);
+							double t2 = exp(-t1);
+							double c0 = log(cg * (1.0 - t2) + t2);
+							double t3 = exp(xx0 - t1);
+							double t6 = exp(-t1);
+							double c1 = -t3 * (cg - 1.0) / (-cg + cg * t6 - t6);
+							t2 = t1;
+							t1 = SQR(cg);
+							double t4 = exp(xx0 - t2);
+							double t8 = exp(xx0 - 2.0 * t2);
+							double t14 = exp(2.0 * xx0 - t2);
+							double t19 = exp(-t2);
+							double t22 = SQR(-cg + cg * t19 - t19);
+							double c2 =
+								-(-t1 * t4 + t1 * t8 - 2.0 * cg * t8 + t1 * t14 + cg * t4 + t8 - cg * t14) / t22;
+							logll[i] = c0 + dx * (c1 + 0.5 * c2 * dx);
+						}
+					}
 				} else {
 					for (int i = 0; i < m; i++) {
 						double phi = PREDICTOR_INVERSE_LINK(x[i] + off);
@@ -3127,9 +3213,18 @@ int loglikelihood_occupancy(int thread_id, double *__restrict logll, double *__r
 					}
 				}
 			} else {
-				for (int i = 0; i < m; i++) {
-					double phi = PREDICTOR_INVERSE_LINK(x[i] + off);
-					logll[i] = logll0 + LOG_p(phi);
+				// not yzero
+				if (PREDICTOR_LINK_EQ(link_cloglog)) {
+#pragma omp simd
+					for (int i = 0; i < m; i++) {
+						logll[i] = logll0 + link_log_invcloglog(PREDICTOR_INVERSE_IDENTITY_LINK(x[i] + off));
+					}
+				} else {
+#pragma omp simd
+					for (int i = 0; i < m; i++) {
+						double phi = PREDICTOR_INVERSE_LINK(x[i] + off);
+						logll[i] = logll0 + LOG_p(phi);
+					}
 				}
 			}
 		}
@@ -3294,9 +3389,11 @@ int loglikelihood_binomialmix(int thread_id, double *__restrict logll, double *_
 	}
 
 	Data_section_tp *ds = (Data_section_tp *) arg;
+	int mm = ds->data_observations.binmix_m;
+	int nbeta = 2 * mm + 1;
 	double y = ds->data_observations.y[idx];
 	double *dat = ds->data_observations.binmix_dat[idx];
-	double n = dat[BINOMIALMIX_NBETA + 4];		       /* z_1, ..., z_11, w_1, w_2 */
+	double n = dat[nbeta + 4];		      
 	double ny = n - y;
 
 	if (ISZERO(y) && ISZERO(n)) {
@@ -3316,24 +3413,24 @@ int loglikelihood_binomialmix(int thread_id, double *__restrict logll, double *_
 	double normc = G_norm_const[idx];
 	LINK_INIT;
 
-	double beta[BINOMIALMIX_NBETA];
-	for (int i = 0; i < BINOMIALMIX_NBETA; i++) {
+	double beta[nbeta];
+	for (int i = 0; i < nbeta; i++) {
 		beta[i] = ds->data_observations.binmix_beta[i][thread_id][0];
 	}
-	double beta9 = beta[BINOMIALMIX_NBETA - 1];
+	double beta_common = beta[nbeta - 1];
 
-	int nbeta2 = BINOMIALMIX_NBETA / 2L;
-	double p1_intern = beta9 * dat[BINOMIALMIX_NBETA - 1] + GMRFLib_ddot(nbeta2, beta, dat);
-	double p2_intern = beta9 * dat[BINOMIALMIX_NBETA + 0] + GMRFLib_ddot(nbeta2, beta + nbeta2, dat + nbeta2);
+	int nbeta2 = nbeta / 2L;
+	double p1_intern = beta_common * dat[nbeta - 1] + (nbeta2 > 0 ? GMRFLib_ddot(nbeta2, beta, dat) : 0.0);
+	double p2_intern = beta_common * dat[nbeta + 0] + (nbeta2 > 0 ? GMRFLib_ddot(nbeta2, beta + nbeta2, dat + nbeta2) : 0.0);
 	double p1 = PREDICTOR_INVERSE_LINK_NO_SCALE(p1_intern);
 	double p2 = PREDICTOR_INVERSE_LINK_NO_SCALE(p2_intern);
-	double w1 = dat[BINOMIALMIX_NBETA + 2];
-	double w2 = dat[BINOMIALMIX_NBETA + 3];
+	double w1 = dat[nbeta + 2];
+	double w2 = dat[nbeta + 3];
 	double p12 = w1 * p1 + w2 * p2;
 	double w3 = 1.0 - (w1 + w2);
 	assert(w3 >= 0 && w3 <= 1.0);
 
-	double offset = off + beta9 * dat[BINOMIALMIX_NBETA + 1];
+	double offset = off + beta_common * dat[nbeta + 1];
 	if (m > 0) {
 		if (ISZERO(y)) {
 #pragma omp simd
@@ -5761,9 +5858,9 @@ int loglikelihood_mix_gaussian(int thread_id, double *__restrict logll, double *
 
 int loglikelihood_mix_core(int thread_id, double *__restrict logll, double *__restrict x, int m, int idx, double *x_vec, double *y_cdf, void *arg,
 			   int (*func_quadrature)(int, double **, double **, int *, void *arg),
-			   int(*func_simpson)(int, double **, double **, int *, void *arg), char **arg_str)
+			   int (*func_simpson)(int, double **, double **, int *, void *arg), char **arg_str)
 {
-	Data_section_tp *ds =(Data_section_tp *) arg;
+	Data_section_tp *ds = (Data_section_tp *) arg;
 	if (m == 0) {
 		if (arg) {
 			return (ds->mix_loglikelihood(thread_id, NULL, NULL, 0, 0, NULL, NULL, arg, arg_str));